Backend/requirements.txt
flask
gunicorn
yt-dlp
flask-cors
requests
python-dotenv
